Modularize scripture schema
